adding camera (no scan yet)
