+++ client/babel.config.js
@@ -1,4 +1,7 @@
 module.exports = {
   presets: ["babel-preset-expo"],
-  plugins: ["react-native-reanimated/plugin"],
+  plugins: [
+    ["react-native-reanimated/plugin"],
+    ["react-native-worklets-core/plugin"],
+  ],
 };
