+++ client/hooks/useCamera.ts
@@ -1,10 +1,11 @@
 import { Platform } from "react-native";
 
+// Définition par défaut pour le Web
 export let useCameraDevices: any = () => ({ back: null });
 export let useFrameProcessor: any = () => undefined;
 export let useScanBarcodes: any = () => [
   (frame: any) => {
-    "worklet";
+    "worklet"; // Cette ligne marque la fonction comme worklet
     return [];
   },
   [],
@@ -15,11 +16,11 @@ if (Platform.OS !== "web") {
   useCameraDevices = vision.useCameraDevices;
   useFrameProcessor = vision.useFrameProcessor;
 
-  // Simulation d’un scan désactivé, mais worklet-compatible
+  // Déclaration correcte de la fonction worklet
   useScanBarcodes = () => [
     (frame: any) => {
-      "worklet";
-      return [];
+      "worklet"; // Cette ligne marque la fonction comme worklet
+      return []; // Simulation d'un scan désactivé ou vide
     },
     [],
   ];
