+++ client/package.json
@@ -54,11 +54,12 @@
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
     "react-native-worklets-core": "^1.5.0",
-    "react-webcam": "^7.2.0"
+    "react-webcam": "^7.2.0",
+    "vision-camera-code-scanner": "^0.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@types/jest": "^29.5.12",
+    "@types/jest": "^29.5.14",
     "@types/react": "~18.3.12",
     "@types/react-test-renderer": "^18.3.0",
     "jest": "~29.7.0",
