+++ client/tsconfig.json
@@ -6,7 +6,8 @@
       "@/*": ["./*"],
       "@env": ["./env.d.ts"]
     },
-    "module": "esnext"
+    "module": "esnext",
+    "types": ["jest", "node"]
   },
   "include": [
     "**/*.ts",
